requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cycler==0.10.0
Deprecated==1.2.13
idna==3.2
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
